editor/libeditor/tests/test_bug1399722.html

<!DOCTYPE html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1399722
-->
<html>
<head>
  <title>Test for Bug 1399722</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1399722">Mozilla Bug 1399722</a>
<p id="display"></p>
<div id="content" style="display: none;">

</div>

<input spellcheck="true" onkeypress="if (event.key=='Enter')this.value='';">
<pre id="test">

<script class="testbody" type="application/javascript">
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  let elm = document.querySelector("input");

  elm.focus();
  synthesizeKey("A", {});
  synthesizeKey("B", {});
  synthesizeKey("KEY_Enter", {});
  synthesizeKey("C", {});
  synthesizeKey("D", {});
  is(elm.value, "CD", "Can type into the textbox successfully after the onkeypress handler deleting the value");

  SimpleTest.finish();
});
</script>
</pre>
</body>
</html>
